core: replace with new api code
