+++ Blasphemous.Debug.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.8.34408.163
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blasphemous.Debug", "Blasphemous.Debug\Blasphemous.Debug.csproj", "{27719701-1BE1-43FE-9023-E1E8B7936F21}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{27719701-1BE1-43FE-9023-E1E8B7936F21}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{27719701-1BE1-43FE-9023-E1E8B7936F21}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{27719701-1BE1-43FE-9023-E1E8B7936F21}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{27719701-1BE1-43FE-9023-E1E8B7936F21}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A5ADD168-3C12-4B7A-96FE-B9DB54651D07}
+	EndGlobalSection
+EndGlobal
