+++ Blasphemous.Debug/IModule.cs
@@ -1,0 +1,11 @@
+﻿
+namespace Blasphemous.Debug;
+
+public interface IModule
+{
+    public void OnLevelLoaded();
+
+    public void OnLevelUnloaded();
+
+    public void Update();
+}
