--- DebugMod.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31605.320
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DebugMod", "DebugMod\DebugMod.csproj", "{40972432-09C3-4F76-8A5B-47212CA37CC7}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{40972432-09C3-4F76-8A5B-47212CA37CC7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{40972432-09C3-4F76-8A5B-47212CA37CC7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{40972432-09C3-4F76-8A5B-47212CA37CC7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{40972432-09C3-4F76-8A5B-47212CA37CC7}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F7D8C6A4-822C-49AE-A0E6-2B370F7B30D2}
-	EndGlobalSection
-EndGlobal
