--- DebugMod/Main.cs
@@ -1,17 +1,0 @@
-﻿using BepInEx;
-
-namespace DebugMod
-{
-    [BepInPlugin(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION)]
-    [BepInDependency("com.damocles.blasphemous.modding-api", "1.4.0")]
-    [BepInProcess("Blasphemous.exe")]
-    public class Main : BaseUnityPlugin
-    {
-        public static Debugger Debugger { get; private set; }
-
-        private void Start()
-        {
-            Debugger = new Debugger(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION);
-        }
-    }
-}
